Arachne 1717 1.1.0 release

+++ src/main/java/com/odysseusinc/arachne/datanode/service/impl/DataSourceServiceImpl.java
@@ -26,6 +26,7 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.google.common.base.Preconditions;
 import com.odysseusinc.arachne.commons.api.v1.dto.CommonHealthStatus;
+import com.odysseusinc.arachne.datanode.exception.IllegalOperationException;
 import com.odysseusinc.arachne.datanode.exception.NotExistException;
 import com.odysseusinc.arachne.datanode.model.datanode.DataNode;
 import com.odysseusinc.arachne.datanode.model.datasource.DataSource;
@@ -33,6 +34,11 @@ import com.odysseusinc.arachne.datanode.model.user.User;
 import com.odysseusinc.arachne.datanode.repository.DataSourceRepository;
 import com.odysseusinc.arachne.datanode.service.DataNodeService;
 import com.odysseusinc.arachne.datanode.service.DataSourceService;
+import java.util.HashMap;
+import java.util.Map;
+import javax.annotation.PostConstruct;
+import org.springframework.data.domain.Sort;
+
 import com.odysseusinc.arachne.execution_engine_common.api.v1.dto.DBMSType;
 import java.util.List;
 import java.util.Objects;
@@ -51,6 +57,7 @@ public class DataSourceServiceImpl implements DataSourceService {
 
     private final DataSourceRepository dataSourceRepository;
     private final DataNodeService dataNodeService;
+    private final Map<String, String> dsSortPath = new HashMap<>();
 
     @Autowired
     public DataSourceServiceImpl(DataSourceRepository dataSourceRepository,
@@ -58,6 +65,17 @@ public class DataSourceServiceImpl implements DataSourceService {
 
         this.dataSourceRepository = dataSourceRepository;
         this.dataNodeService = dataNodeService;
+    }
+
+    @PostConstruct
+    private void init() {
+
+        this.dsSortPath.put("name", "name");
+        this.dsSortPath.put("dbmsType", "type");
+        this.dsSortPath.put("connectionString", "connectionString");
+        this.dsSortPath.put("cdmSchema", "cdmSchema");
+        this.dsSortPath.put("modelType", "modelType");
+        this.dsSortPath.put("isRegistered", "registred");
     }
 
     @Transactional(rollbackFor = Exception.class)
@@ -83,6 +101,12 @@ public class DataSourceServiceImpl implements DataSourceService {
     }
 
     @Override
+    public List<DataSource> findAll(String sortBy, Boolean sortAsc) {
+
+        return dataSourceRepository.findAll(getSort(sortBy, sortAsc));
+    }
+
+    @Override
     public List<DataSource> findAllRegistered() {
 
         return dataSourceRepository.findAllRegistered();
@@ -92,6 +116,12 @@ public class DataSourceServiceImpl implements DataSourceService {
     public void delete(Long id) {
 
         checkNotNull(id, "given data source surrogate id is blank ");
+        final DataSource dataSource = getById(id);
+        if (dataSource.getRegistred()) {
+            final String message
+                    = String.format("Can not delete registered DataSource with id='%s'. Unregister it first", id);
+            throw new IllegalOperationException(message);
+        }
         dataSourceRepository.delete(id);
     }
 
@@ -207,5 +237,14 @@ public class DataSourceServiceImpl implements DataSourceService {
 
         dataSource.setRegistred(centralId != null);
         return dataSourceRepository.save(dataSource);
+    }
+
+    protected final Sort getSort(String sortBy, Boolean sortAsc) {
+
+        String defaultSort = "name";
+        return new Sort(
+                sortAsc == null || sortAsc ? Sort.Direction.ASC : Sort.Direction.DESC,
+                dsSortPath.getOrDefault(sortBy, defaultSort)
+        );
     }
 }
